# backend/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lost & Found Pet Network API",
    "description": "A comprehensive API for managing lost and found pets. Users can post missing or found pets, search for pets, contact owners, and report issues. Admins can moderate content and manage users.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@petadoption.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:5001",
      "description": "Alternative development server"
    }
  ],
...
  "paths": {
...
    "/api/pets": {
      "get": {
        "tags": ["Pets"],
        "summary": "Get all pets",
        "description": "Retrieve pets with optional filtering and pagination",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by pet status",
            "schema": {
              "type": "string",
              "enum": ["missing", "found", "reunited"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by pet type",
            "schema": {
              "type": "string",
              "enum": ["dog", "cat", "bird", "rabbit", "hamster", "fish", "other"]
            }
          },
          {
            "name": "breed",
            "in": "query",
            "description": "Search by breed",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "color",
            "in": "query",
            "description": "Search by color",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "description": "Filter by gender",
            "schema": {
              "type": "string",
              "enum": ["male", "female", "unknown"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pets retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Pets"],
        "summary": "Create pet post",
        "description": "Create a new pet post (missing or found)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pet post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/pets": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all pets",
        "description": "Retrieve all pets with pagination and filtering",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by pet status",
            "schema": {
              "type": "string",
              "enum": ["missing", "found", "reunited"]
            }
          },
          {
            "name": "isApproved",
            "in": "query",
            "description": "Filter by approval status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pets retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "maxLength": 100
          },
          "city": {
            "type": "string",
            "maxLength": 50
          },
          "state": {
            "type": "string",
            "maxLength": 50
          },
          "zipCode": {
            "type": "string",
            "maxLength": 10
          }
        }
      },
      "PetCreate": {
        "type": "object",
        "required": ["name", "type", "breed", "color", "gender", "status", "lastSeenLocation", "lastSeenDate"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "type": {
            "type": "string",
            "enum": ["dog", "cat", "bird", "rabbit", "hamster", "fish", "other"]
          },
          "breed": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "unknown"]
          },
          "age": {
            "type": "number",
            "minimum": 0,
            "maximum": 30
          },
          "ageUnit": {
            "type": "string",
            "enum": ["days", "weeks", "months", "years"]
          },
          "weight": {
            "type": "number",
            "minimum": 0
          },
          "weightUnit": {
            "type": "string",
            "enum": ["kg", "lbs"]
          },
          "status": {
            "type": "string",
            "enum": ["missing", "found", "reunited"]
          },
          "lastSeenLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "lastSeenDate": {
            "type": "string",
            "format": "date-time"
          },
          "additionalNotes": {
            "type": "string",
            "maxLength": 1000
          },
          "microchipNumber": {
            "type": "string",
            "maxLength": 50
          },
          "collar": {
            "$ref": "#/components/schemas/Collar"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 50
            }
          }
        }
      },
...
      "Location": {
        "type": "object",
        "required": ["address", "coordinates"],
        "properties": {
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                },
                "minItems": 2,
                "maxItems": 2
              }
            }
          },
          "city": {
            "type": "string",
            "maxLength": 50
          },
          "state": {
            "type": "string",
            "maxLength": 50
          },
          "zipCode": {
            "type": "string",
            "maxLength": 10
          }
        }
      },
      "Collar": {
        "type": "object",
        "properties": {
          "hasCollar": {
            "type": "boolean"
          },
          "color": {
            "type": "string",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "profilePicture": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["user", "admin", "shelter", "vet"]
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PetResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "pet": {
                "$ref": "#/components/schemas/Pet"
              }
            }
          }
        }
      },
      "Pet": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "breed": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "age": {
            "type": "number"
          },
          "ageUnit": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "weightUnit": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastSeenLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "lastSeenDate": {
            "type": "string",
            "format": "date-time"
          },
          "additionalNotes": {
            "type": "string"
          },
          "microchipNumber": {
            "type": "string"
          },
          "collar": {
            "$ref": "#/components/schemas/Collar"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isApproved": {
            "type": "boolean"
          },
          "views": {
            "type": "number"
          },
          "contactCount": {
            "type": "number"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PetsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "pets": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Pet"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              }
            }
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}